requirements.txt
contourpy==1.3.0
cycler==0.12.1
flatbuffers==24.12.23
fonttools==4.55.3
google-auth-oauthlib==1.0.0
grpcio==1.68.1
importlib-resources==6.4.5
jax==0.4.30
jaxlib==0.4.30
keras==2.13.1
keras-preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
matplotlib==3.9.4
mediapipe==0.10.5
ml-dtypes==0.5.0
mouseinfo==0.1.3
numpy==1.23.5
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
pillow==11.0.0
protobuf==3.20.3
pyautogui==0.9.54
pygetwindow==0.0.9
pymsgbox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
pyrect==0.2.0
pyscreeze==1.0.1
pytweening==1.2.0
sentencepiece==0.2.0
sounddevice==0.5.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
typing-extensions==4.5.0
wrapt==1.14.1
